refactor(HomepageHero): move to general-widgets
