+++ src/components/general-widgets/HomepageHero/component.spec.js
