+++ src/components/general-widgets/HomepageHero/mock-data/livingstoneLogo.png
