+++ src/components/general-widgets/HomepageHero/mock-data/mock-data.js
