+++ src/components/general-widgets/HomepageHero/utils/getSearchBarMarkup.js
