+++ src/components/general-widgets/HomepageHero/utils/getSearchBarMarkup.spec.js
